+++ src/constants/atomKeys.ts
@@ -1,0 +1,3 @@
+export const ATOM_KEYS = {
+  MOIT_REGISTER_SELECT_BOTTOM_SHEET_STATUS: 'SELECT_BOTTOM_SHEET',
+};
