--- src/domain/moit/constants/data.ts
@@ -1,7 +1,0 @@
-const now = new Date();
-
-export const INITIAL_DATE: DateParams = {
-  y: 0,
-  m: now.getMonth(),
-  d: now.getDate(),
-};
