+++ src/screens/MoitRegisterScreen/components/DateZone.tsx
