+++ src/screens/MoitRegisterScreen/hooks/useSelectTime.ts
@@ -1,10 +1,7 @@
 import { UseSelectScroller } from '@components/SelectScroller/SelectScroller.hooks';
 import { useSelectScroller } from '@components/SelectScroller';
-import { CreateMoitRegisterTime } from '../components/TimeSelectBottomSheet';
-import {
-  SELECT_CONTENT_HEIGHT,
-  SELECT_TIME_MINUTE_INTERVAL,
-} from '../constants';
+import { CreateMoitRegisterTime } from '../components/TimeSelectScreen';
+import { SELECT_CONTENT_HEIGHT, SELECT_TIME_MINUTE_INTERVAL } from '../consts';
 
 interface UseSelectTimeProps {
   hour: UseSelectScroller;
