+++ src/screens/MoitRegisterScreen/styled.ts
